move: component split to specific folder
