--- src/app/menu/loading.tsx
@@ -1,7 +1,0 @@
-export default function Loading() {
-  return (
-    <div>
-      <p>리스트를 불러오는 중입니다...</p>
-    </div>
-  );
-}
