+++ src/app/menu/page.module.css
@@ -1,6 +1,7 @@
 .client {
   height: calc(100% - var(--header-height, 50px));
   --name-section-height: 85px;
+  --category-height: 4.5rem;
 }
 
 /** name section */
@@ -12,61 +13,9 @@
   background-color: beige;
 }
 
-.modal_container {
-  position: relative;
-  width: 100%;
-  height: 300px;
-  background-color: #fff;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.modal_container button.close {
-  position: absolute;
-  top: 1rem;
-  right: 1rem;
-}
-
-.modal_container form {
-  width: 80%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.modal_container form input {
-  height: 2rem;
-  width: 100%;
-}
-.modal_container form button {
-  margin-top: 1rem;
-  height: 2rem;
-}
-/** 메뉴 선택 모달 */
-.modal_container .menu-column {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  margin-bottom: 1rem;
-}
-
-.modal_container .menu-column label {
-  font-size: 1.5rem;
-  margin-bottom: 0.5rem;
-}
-.modal_container .menu-column select {
-  width: 100%;
-  height: 2rem;
-}
-
 .menu_container {
-  height: calc(100% - var(--name-section-height));
+  height: calc(100% - var(--name-section-height) - var(--category-height));
   overflow: scroll;
-  --category-height: 4.5rem;
   --footer-height: 5rem;
 }
 
@@ -97,7 +46,7 @@
 }
 
 .menu {
-  min-height: calc(100% - var(--category-height) - var(--footer-height) - 4rem);
+  min-height: calc(100% - var(--footer-height) - 4rem);
   margin-bottom: 3rem;
 }
